+++ app/components/CredentialCard/DefaultCredentialCard/DefaultCredentialCard.tsx
@@ -4,15 +4,12 @@ import { View, Text, Image, Linking } from 'react-native';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 
 import type { CredentialRecordRaw } from '../../model/credential';
-import { theme } from '../../styles';
-import styles from './DefaultCredentialCard.styles';
+import { theme } from '../../../styles';
+import styles from '../CredentialCard.styles';
+import type { CredentialCardProps } from '../CredentialCard.d';
 
 const NO_URL = 'None';
 const DATE_FORMAT = 'MMM D, YYYY';
-
-type DefaultCredentialCardProps = {
-  rawCredentialRecord: CredentialRecordRaw;
-};
 
 
 export default function DefaultCredentialCard({ rawCredentialRecord }: DefaultCredentialCardProps): JSX.Element {
